Tidied modules structure.
